Alle funktionaliteter er lavet nu uden logging, dog er der en lille fejl. Den catalogId der kommer med i auction når du opretter, findes ikke når du prøver at se catalog objektet. ved ikke om vi bare skal hardcode nogle faste catalogs?

+++ AuctionServiceAPI/Models/Auction.cs
@@ -9,7 +9,7 @@ public class Auction
     public Guid CatalogId { get; set; }  // FK
     public List<BidDTO> BidHistory { get; set; } = new(); // FK - Historik over bud
 
-    public Decimal MinPrice { get; set; }    // Det nuværende højeste bud - FK
+    public double MinPrice { get; set; }    // Det nuværende højeste bud - FK
     
     public BidDTO? CurrentBid { get; set; }  // Det nuværende højeste bud - FK
